fix(channel-search): split decode helper into dep-free module
CI unit-tests regression on 600f721e/e8e8e529: importing
decodeServerEscapedHighlight from snippetContent.tsx into
SearchResultMapper.ts pulled the App component tree (react-virtuoso and
friends) transitively into pure-data test files that don't mock those
ESM modules, so vitest collection failed with:

  Cannot find module '.../react-virtuoso/.../react/jsx-runtime'

Four otherwise-unrelated test files (SearchService.test.ts,
SearchService.drive.test.ts, filterPanel.render.test.tsx,
filterPanel.dateHelpers.test.ts) reported "0 test" because their
imports transitively touched SearchResultMapper -> snippetContent
-> App tree. Local vitest reproduces; CI's turbo run test:coverage
exits 1 on it.

Move decodeServerEscapedHighlight into its own module,
Components/ChannelSearch/decodeHighlight.ts, with zero React or App
imports. snippetContent.tsx re-exports it so the public API is stable
and existing test imports keep working. SearchResultMapper switches
its import to the pure module — cuts the App-tree edge on the pure-
data path.

Tests: SearchService.test.ts + SearchService.drive.test.ts +
filterPanel.render.test.tsx + filterPanel.dateHelpers.test.ts +
snippetContent.test.tsx + apiAdapter.test.ts +
ChannelSearchResults.helpers.test.tsx = 7 files / 57 tests, all pass
locally. Prettier clean.
